이전에 사용했던 로직들을 우선 그대로 구조체 단위로 나눔
